requirements.txt
aiogram==2.9.2
aiohttp==3.6.2
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.2
gunicorn==20.0.4
hiredis==1.1.0
idna==2.10
multidict==4.7.6
psycopg2==2.8.6
python-dotenv==0.14.0
pytz==2020.1
sqlparse==0.4.1
whitenoise==5.2.0
yarl==1.6.0